// MyNotesFront/angular-my-notes/src/app/models/user.ts
export interface User{
  email:string|null|undefined;
  firstName: string |null |undefined;
  lastName: string |null |undefined;
  username: string |null |undefined;
  password:string |null |undefined;
  passwordConfirmation:string |null |undefined;
  address: string |null |undefined;
  phoneNumber: string |null |undefined;
}